working on dailystandup.js
