+++ b2b-app/imports/ui/admin/standup-notes/edit-team.js
@@ -1,10 +1,74 @@
-import { Button } from '@material-ui/core'
-import { DataGrid } from '@mui/x-data-grid'
-import { rows, columns } from './edit-team.stories.js'
-import IconButton from '@material-ui/core/IconButton'
 import React from 'react'
+import { Button } from '@material-ui/core'
+import IconButton from '@material-ui/core/IconButton'
+import DataGrid from 'react-data-grid'
 
-const editTeam = () => {
+const columns = [
+  {
+    field: 'firstName',
+    headerName: 'First name',
+    width: 150,
+    editable: true,
+  },
+  {
+    field: 'handle',
+    headerName: 'Handle',
+    width: 150,
+    editable: true,
+  },
+  {
+    field: 'joined',
+    headerName: 'Joined',
+    type: 'date',
+    width: 110,
+    editable: true,
+  },
+  {
+    field: 'active',
+    headerName: 'Active',
+    width: 1,
+    editable: true,
+  },
+  {
+    field: 'delete',
+    headerName: 'Delete',
+    width: 1,
+    editable: true,
+  },
+]
+
+const rows = [
+  {
+    firstName: 'Mike King',
+    handle: 'mikkel',
+    joined: '1/02/2021',
+    active: 'Y',
+    delete: <Button>x</Button>,
+  },
+  {
+    firstName: 'Patrick Carmelt',
+    handle: 'pato',
+    joined: '1/2/2021',
+    active: 'Y',
+    delete: <Button>x</Button>,
+  },
+  {
+    firstName: 'Chris Tri',
+    handle: 'ct',
+    joined: '1/12/2021',
+    active: 'N',
+    delete: <Button>x</Button>,
+  },
+  {
+    firstName: 'Minh Ngyuen',
+    handle: 'minster',
+    joined: '1/12/2021',
+    active: 'Y',
+    delete: <Button>x</Button>,
+  },
+]
+
+const EditTeam = () => {
   return (
     <div>
       <h1>Edit Team </h1>
@@ -22,12 +86,12 @@ const editTeam = () => {
         Team name
       </Typography>
 
-      <Autocomplete
+      {/* <Autocomplete
         id="team-input"
         freesolo
         options={['The dream team', 'Peaky Blinders', 'Gary and the pacemakers']}
         renderInput={(params) => (
-          <TextField {...params} label="member name" margin="normal" variant="outlined" />
+          <TextField {...params} label="team name" margin="normal" variant="outlined" />
         )}
         style={{
           width: '180px',
@@ -53,11 +117,11 @@ const editTeam = () => {
         }}
       >
         <AddIcon />
-      </IconButton>
+      </IconButton> */}
 
       <DataGrid
-        rows={rows}
-        columns={columns}
+        rows={[]}
+        columns={[]}
         pageSize={5}
         rowsPerPageOptions={[5]}
         checkboxSelection
@@ -67,4 +131,4 @@ const editTeam = () => {
   )
 }
 
-export default editTeam
+export default EditTeam
