+++ b2b-app/imports/ui/admin/standup-notes/edit-team.mdx
@@ -2,12 +2,12 @@
 
 import { Meta, Story } from '@storybook/addon-docs'
 
-import editTeam from './edit-team'
+import EditTeam from './edit-team'
 
-<Meta title="editTeam/demo" component={editTeam} />
+<Meta title="EditTeam/demo" component={EditTeam} />
 
-# editTeam
+# EditTeam
 
 <Story name="Primary">
-  <editTeam />
+  <EditTeam />
 </Story>
