+++ b2b-app/imports/ui/pages/daily-standup.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const DailyStandupPage = () => {
+  return <div>Daily standup</div>
+}
+
+export default DailyStandupPage
